third_party/blink/web_tests/external/wpt/css/selectors/invalidation/has-in-adjacent-position-expected.txt

This is a testharness.js-based test.
Found 71 tests; 47 PASS, 24 FAIL, 0 TIMEOUT, 0 NOTRUN.
PASS Initial color
PASS add .test to previous_sibling
PASS remove .test from previous_sibling
FAIL add .test to previous_sibling_child assert_equals: expected "rgb(0, 128, 0)" but got "rgb(128, 128, 128)"
PASS remove .test from previous_sibling_child
FAIL add .test to previous_sibling_descendant assert_equals: expected "rgb(255, 0, 0)" but got "rgb(128, 128, 128)"
PASS remove .test from previous_sibling_descendant
FAIL add .test to subject assert_equals: expected "rgb(0, 0, 255)" but got "rgb(128, 128, 128)"
PASS remove .test from subject
FAIL add .test to next_sibling assert_equals: expected "rgb(255, 255, 0)" but got "rgb(128, 128, 128)"
PASS remove .test from next_sibling
FAIL add .test to next_sibling_child assert_equals: expected "rgb(128, 0, 128)" but got "rgb(128, 128, 128)"
PASS remove .test from next_sibling_child
FAIL add .test to next_sibling_descendant assert_equals: expected "rgb(128, 0, 128)" but got "rgb(128, 128, 128)"
PASS remove .test from next_sibling_descendant
PASS insert element div.test before previous_sibling
PASS remove element div.test before previous_sibling
FAIL insert element div.test before previous_sibling_child assert_equals: expected "rgb(0, 128, 0)" but got "rgb(128, 128, 128)"
PASS remove element div.test before previous_sibling_child
FAIL insert element div.test before previous_sibling_descendant assert_equals: expected "rgb(255, 0, 0)" but got "rgb(128, 128, 128)"
PASS remove element div.test before previous_sibling_descendant
PASS insert element div.test before subject
PASS remove element div.test before subject
PASS insert element div.test before next_sibling
PASS remove element div.test before next_sibling
FAIL insert element div.test before next_sibling_child assert_equals: expected "rgb(128, 0, 128)" but got "rgb(128, 128, 128)"
PASS remove element div.test before next_sibling_child
FAIL insert element div.test before next_sibling_descendant assert_equals: expected "rgb(128, 0, 128)" but got "rgb(128, 128, 128)"
PASS remove element div.test before next_sibling_descendant
PASS insert element div.test after previous_sibling
PASS remove element div.test after previous_sibling
FAIL insert element div.test after previous_sibling_child assert_equals: expected "rgb(0, 128, 0)" but got "rgb(128, 128, 128)"
PASS remove element div.test after previous_sibling_child
FAIL insert element div.test after previous_sibling_descendant assert_equals: expected "rgb(255, 0, 0)" but got "rgb(128, 128, 128)"
PASS remove element div.test after previous_sibling_descendant
PASS insert element div.test after subject
PASS remove element div.test after subject
FAIL insert element div.test after next_sibling assert_equals: expected "rgb(255, 255, 0)" but got "rgb(128, 128, 128)"
PASS remove element div.test after next_sibling
FAIL insert element div.test after next_sibling_child assert_equals: expected "rgb(128, 0, 128)" but got "rgb(128, 128, 128)"
PASS remove element div.test after next_sibling_child
FAIL insert element div.test after next_sibling_descendant assert_equals: expected "rgb(128, 0, 128)" but got "rgb(128, 128, 128)"
PASS remove element div.test after next_sibling_descendant
PASS insert tree div>div.test before previous_sibling
PASS remove tree div>div.test before previous_sibling
FAIL insert tree div>div.test before previous_sibling_child assert_equals: expected "rgb(255, 0, 0)" but got "rgb(128, 128, 128)"
PASS remove tree div>div.test before previous_sibling_child
FAIL insert tree div>div.test before previous_sibling_descendant assert_equals: expected "rgb(255, 0, 0)" but got "rgb(128, 128, 128)"
PASS remove tree div>div.test before previous_sibling_descendant
PASS insert tree div>div.test before subject
PASS remove tree div>div.test before subject
PASS insert tree div>div.test before next_sibling
PASS remove tree div>div.test before next_sibling
FAIL insert tree div>div.test before next_sibling_child assert_equals: expected "rgb(128, 0, 128)" but got "rgb(128, 128, 128)"
PASS remove tree div>div.test before next_sibling_child
FAIL insert tree div>div.test before next_sibling_descendant assert_equals: expected "rgb(128, 0, 128)" but got "rgb(128, 128, 128)"
PASS remove tree div>div.test before next_sibling_descendant
PASS insert tree div>div.test after previous_sibling
PASS remove tree div>div.test after previous_sibling
FAIL insert tree div>div.test after previous_sibling_child assert_equals: expected "rgb(255, 0, 0)" but got "rgb(128, 128, 128)"
PASS remove tree div>div.test after previous_sibling_child
FAIL insert tree div>div.test after previous_sibling_descendant assert_equals: expected "rgb(255, 0, 0)" but got "rgb(128, 128, 128)"
PASS remove tree div>div.test after previous_sibling_descendant
PASS insert tree div>div.test after subject
PASS remove tree div>div.test after subject
FAIL insert tree div>div.test after next_sibling assert_equals: expected "rgb(128, 0, 128)" but got "rgb(128, 128, 128)"
PASS remove tree div>div.test after next_sibling
FAIL insert tree div>div.test after next_sibling_child assert_equals: expected "rgb(128, 0, 128)" but got "rgb(128, 128, 128)"
PASS remove tree div>div.test after next_sibling_child
FAIL insert tree div>div.test after next_sibling_descendant assert_equals: expected "rgb(128, 0, 128)" but got "rgb(128, 128, 128)"
PASS remove tree div>div.test after next_sibling_descendant
Harness: the test ran to completion.
